Initial addition of MongooseSocket class
MongooseSocket will implement the basic socket management code that all the other protocol classes duplicate

+++ examples/simple_mqtt_client/src/simple_mqtt_client.cpp
@@ -128,7 +128,7 @@ void setup()
 
   snprintf(clientId, sizeof(clientId), "mg-%llx", deviceId);
   client.onMessage([](MongooseString topic, MongooseString payload) {
-    DBUGF("%.*s: %.*s", topic.length(), (const char *)topic, payload.length(), (const char *)payload);
+    DBUGF("%.*s: %.*s", topic.length(), (const char *)topic, (int)payload.length(), (const char *)payload);
     client.publish("/test", payload);
   });
   client.onError([](uint8_t err) {
